backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25
alembic==1.13.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.2

# Embeddings and ML
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.1.2

# LLM Integration
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Development and Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
